FCT Tiroir de navigation, amilioration du listView et création de tâche http

+++ tp1_flutter/lib/transfert.dart
@@ -97,7 +97,7 @@ class GetTasksResponse {
   int id = 0;
   String name = '';
   int percentageDone = 0;
-  double percentageTimeSpent = 0 as double;
+  double percentageTimeSpent = 0.0;
   DateTime deadline = DateTime.now();
 
   factory GetTasksResponse.fromJson(Map<String, dynamic> json) =>
